refactor: ui 層を presentation 層にリネーム
- src/ui/ → src/presentation/ にディレクトリをリネーム
- main.tsx の import パスを更新

Closes #12
