+++ src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import { App } from './ui/App';
+import { App } from './presentation/App';
 import './index.css';
 
 const rootElement = document.getElementById('root');
