+++ src/presentation/PlayerCard.tsx
