+++ src/presentation/PointStickDivider.tsx
