+++ src/presentation/controls/CounterInput.tsx
