+++ src/presentation/controls/PlayerSelector.tsx
