+++ src/presentation/dialogs/RiichiDialog.tsx
