+++ src/presentation/dialogs/RyukyokuDialog.tsx
