+++ src/presentation/dialogs/TsumoDialog.tsx
